# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for BPM.",
        "title": "BPM API",
        "contact": {
            "name": "Lewis",
            "email": "lxx0103@yahoo.com"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
...
        "/components": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组件管理"
                ],
                "summary": "组件列表",
                "operationId": "13",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页行数",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "事件ID",
                        "name": "event_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组件编码",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/component.Component"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
        "/components/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组件管理"
                ],
                "summary": "根据ID获取组件",
                "operationId": "15",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "组件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/component.Component"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
...
        "/wx/components": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序接口"
                ],
                "summary": "组件列表",
                "operationId": "87",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页行数",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "事件ID",
                        "name": "event_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组件编码",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ListRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/component.Component"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "component.Component": {
            "type": "object",
            "properties": {
                "component_type": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "default_value": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "json_data": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "patterns": {
                    "type": "string"
                },
                "required": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "updated": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ListRes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {},
                "page_id": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
        "response.SuccessRes": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
    }
}